Implemented smooth movement and hitbox-based collision

+++ game/core/src/main/java/com/marton/roguelike/entity/Player.java
@@ -1,7 +1,7 @@
 package com.marton.roguelike.entity;
 
 public class Player extends Entity{
-    public Player(int x, int y) {
+    public Player(float x, float y) {
         super(x, y);
     }
 }
